style: clear remaining ruff findings (RUF046/RUF005/F601/ARG002/E402)
- sniffer: drop redundant int() around round()
- plugins/__init__: iterable unpacking instead of list()+concat
- oui: remove duplicate dict keys (245a4c, 00037f)
- handshake._arm: drop unused gate/iface/reason params + update callers/tests
- ruff: ignore E402 in tests (pytest.importorskip pattern); strip now-redundant
  per-line noqa directives
ruff check + format clean; 136 tests green.

+++ src/wlan_dumper/core/sniffer.py
@@ -195,7 +195,7 @@ def dissect_packet(pkt: Any, *, now: float | None = None) -> list[Event]:
             wps = _has_wps_ie(pkt.getlayer(Dot11Elt))
             # beacon_interval is in TUs (1 TU = 1.024 ms); convert to ms.
             interval_tu = int(getattr(beacon, "beacon_interval", 0) or 0)
-            interval_ms = int(round(interval_tu * 1.024)) if interval_tu else 0
+            interval_ms = round(interval_tu * 1.024) if interval_tu else 0
             out.append(
                 BeaconSeen(
                     timestamp=ts,
